refactor: shell 폴더를 root로 이름 변경하여 폴더구조 개선
